Direct Port Manipulation

+++ clock-data-types.h
@@ -12,7 +12,7 @@
 enum ClockState { IDLE, RUNNING, TIMEOUT, MENU, DEMO };
 
 typedef struct {
-  int displayValues[TUBE_COUNT];
+  byte displayValues[TUBE_COUNT];
   unsigned long turnLimitMS;
   char label[4];
 } TurnTimerOption;
@@ -22,7 +22,7 @@ typedef struct {
   unsigned long remainingMS;
 } CountdownValues;
 
-const int TURN_TIMER_OPTIONS_COUNT = 13;
+const byte TURN_TIMER_OPTIONS_COUNT = 13;
 const TurnTimerOption TURN_TIMER_OPTIONS[TURN_TIMER_OPTIONS_COUNT] = {
   { { 7, 2, BLANK, BLANK, BLANK, BLANK }, 259201000UL, "72h" },
   { { 4, 8, BLANK, BLANK, BLANK, BLANK }, 172801000UL, "48h" },
@@ -35,7 +35,7 @@ const TurnTimerOption TURN_TIMER_OPTIONS[TURN_TIMER_OPTIONS_COUNT] = {
   { { BLANK, BLANK, 0, 5, BLANK, BLANK }, 300000UL, "5m" },
   { { BLANK, BLANK, 0, 3, BLANK, BLANK }, 180000UL, "3m" },
   { { BLANK, BLANK, 0, 1, BLANK, BLANK }, 60000UL, "1m" },
-  { { BLANK, BLANK, BLANK, BLANK, 1, 0 }, 10000UL, "10s" },     // for testing
+  { { BLANK, BLANK, BLANK, BLANK, 1, 0 }, 10000UL, "10s" },
   { { BLANK, BLANK, BLANK, BLANK, BLANK, 0 }, 0UL, "n0L" }
 };
